Text - Methods added to find characters and retrieve sub-texts.
The only way to split a text in lines and words is retrieve
character per character and query if it's a white space or
a new line character.
Every time a character is requested, Text creates a Character
object with its implementation.
The Find methods retrieve all the positions of the white
space, new line character or any other character within the
given range saving the creation of all Character objects.

[model]       REDWOOD
[binary_type] PDA
[customer]    OPEN
[issue#]      N/A
[problem]     N/A
[cause]       N/A
[solution]    N/A
[team]        N/A
[horizontal_expansion] N/A

Change-Id: Ie6107edfef5db9eb6d54dd84ec9e662dead8a229
Signed-off-by: Victor Cebollada <v.cebollada@samsung.com>