Refactored shader effect implementation and added new hint for culling.
ShaderEffect code was hard to follow, with many methods having the same name
yet performing different functions.

Have refactored the code to use SetProgram() for it's public api, moved and re-wrote
SetPrograms() and SetWrappedProgram() methods from anonymous namespace to private methods
and added 2 new methods called SetProgramImpl() which actually send the program to the shader
object in the scene graph, one of which queries the shader hints for Fixed
or Flexible vertex hint to determine if culling can occur.

Change-Id: Ic2eaf72a17d337dead1b14746d75f30ffd8ce3f3
Signed-off-by: David Steele <david.steele@partner.samsung.com>