Ensured text mesh is centered in text actor.
Problem:
The vertices of the text mesh are not centered, but are offset very
slightly. Also, for italicized text, the mesh is not centered at all.

Solution:
When generating the vertices, calculate min/max, then re-adjust the
vertices so that the center of the bounding box is at the origin (zero).

Change-Id: I1d04cd652fd7c7cfaae2e0f667f12f4023ac30fe
Signed-off-by: David Steele <david.steele@partner.samsung.com>