Changed STEREO_HORIZONTAL view mode to work with landscape orientation
On Stereo horizontal mode, the left and right images are now rendered to be viewed in landscape mode
preserving aspect ratio. Also fixed bug which caused to incorrectly reuse render list when camera was
different to the one used to update the render list before.

TO set STEREO_HORIZONTAL mode, use '--view 1' as a command-line argument.

Change-Id: I7943b31bc6349c2a70244e86b4b0762e6bdb7ea0