Updates to handle context loss and regain
[Problem]  Dali doesn't handle context loss
[Cause]    Context loss code, e.g. ReplaceSurface(), has never been tested,
but is required for new platforms.

[Solution] Allow the application to choose either to handle context
loss itself (i.e.  it should connect to the new stage signals
ContextLost and ContextRegained), or to let Dali handle context loss
( which means it needs to retain bitmap data and mesh data, adding a
memory overhead )

For Tizen, if the application doesn't handle context loss, DALi only
actually needs to retain the mesh data, as it can reload bitmaps from
file based images (No other bitmap type is discardable).

It also triggers a reload of the font atlas, although that should be
relatively quick, as the glyphs have already been cached, so it should
be as quick as resizing texture atlas.

It also triggers a refresh of all render tasks with framebuffers and RefreshOnce.

For controls and applications that delete their render tasks after they have finished
rendering, they will need to listen for ContextRegained in order to re-run them to
recreate the resultant framebuffer.

Added tests for mesh data discard using public meshes and meshes loaded via model.

Change-Id: I0d89f0dee2dfcf79fcabce19c36ac129de077fa8
Signed-off-by: David Steele <david.steele@partner.samsung.com>