Updates to handle context loss and regain
[Problem]  Dali doesn't handle context loss
[Cause]    Context loss code, e.g. ReplaceSurface(), has never been tested,
but is required for new platforms.

[Solution] Allow the application to choose either to handle context loss itself (i.e.
it should connect to the new application signals), or to let Dali handle context
loss ( which means it needs to retain bitmap data and mesh data, adding a memory overhead )

Currently, it triggers the reload of file based images, even though technically it
doesn't need to, as the bitmap data hasn't been discarded, as we still have the filename
and requested attributes available.

It also triggers a reload of the font atlas, although that should be
relatively quick, as the glyphs have already been cached, so it should
be as quick as resizing texture atlas.

Change-Id: I0d89f0dee2dfcf79fcabce19c36ac129de077fa8
Signed-off-by: David Steele <david.steele@partner.samsung.com>