Added hover event in Dali
Hover events are a collection of touch points at a specific moment in time.
When a multi-touch event occurs, each touch point represents the points that
are currently being hovered or the points where a hover has stopped.

Hover event processor processes hover events and emits the Hovered signal
on the hit actor (and its parents).

Change-Id: Ife8e7c1561158e8f98061eb5037ad3b6944846dd