Move ownership of Image requests back to ImageFactory.
Ownership of requests was previously moved to Image, but ImageFactory still
held raw pointers to the requests.

Changed so that both Images and ImageFactory hold smart pointers to the requests,
Image deletion directly calls into ImageFactory to clean up, and ImageFactory
checks if the request ref count decreases to 1, if so, it cleans up the request.

This ensures that ImageFactory is always holding valid Request objects rather
than possibly holding dangling pointers to deleted objects.

Change-Id: I01f6a45c260a7260470da8feea4380aa4fedd670
Signed-off-by: David Steele <david.steele@partner.samsung.com>