Added thread guards to std::map using classes
Problem:  std::maps are being corrupted. Suspect that this is due to
lack of re-entrancy in std::map, std::multimap.

Solution: Guard all public methods of map using classes with a
DALI_ASSERT_ALWAYS that checks that the methods are being called from the
same thread as the constructor.

Updated ShaderFactory, ImageFactory and ResourceClient.

Change-Id: I2311d3287d6a9e8d548aa8fddaf7633d1c28371b
Signed-off-by: David Steele <david.steele@partner.samsung.com>