Added thread guards to std::map using classes
Problem:  std::maps are being corrupted. Suspect that this is due to
lack of re-entrancy in std::map, std::multimap.

Solution: Guard all public methods of map using classes with a new
DALI_THREAD_CHECK macro that checks that the methods are being called from the
same thread as the constructor, and if not, log an error message and optionally
assert.

Updated ShaderFactory, ImageFactory and ResourceClient.

Also updated UpdateMessageQueue to prevent messages being sent to Update thread
if there is no ThreadLocalStorage (e.g. if called from the wrong thread, or
called during shutdown).

Change-Id: I2311d3287d6a9e8d548aa8fddaf7633d1c28371b
Signed-off-by: David Steele <david.steele@partner.samsung.com>