AArch64 inline assembly atomic inc and dec for RefObject ref & unref
[problem] RefObject::Reference() and RefObject::Unreference() showing
up high on profiles and some online discussion of GCC atomic builtins
not being reliable on AArch64.
[cause] GCC builtins use load exclusive / store exclusive instruction
variant with excessive memory ordering semantics suitable for building
a semaphore or similar.
[solution] AArch64 inline assembly implementations of atomic inc and
dec for RefObject ref & unref and using lower-cost load exclusive /
store exclusive variants which are sufficient for the task.
Reduces cost of atomic inc in a synthetic benchmark on a 64 bit ARM
target from 18.8 nanoseconds to 11.8 nanoseconds.
Uses: ldxr & stxr.

Change-Id: I7842eb8ba604c5a61bb0e826a35100f8548f3d1b
Signed-off-by: Andrew Cox <andrew.cox@partner.samsung.com>