(TapGesture) Added min/max taps, removed touches as it's unsupported
Currently, we needed to create two tap-gesture-detectors if we wanted both a single and double tap.
Now we should be able to set the minimum/maximum taps required on each tap-gesture-detector and
just check the noOfTaps member to see how many taps we received.

The touches setter/getter was removed as this is not supported by adaptor. Better to remove this
rather than have unsupported API. Left it in the structs as it could be supported later so that
their sizes do not change when this does happen.

Change-Id: I6112197f867da5a12a2d7dcc20990202a42ca670