Uniform map handling
Updated event implementations of geometry, property buffer, material, sampler and renderer to handle custom properties and default animatable properties.

Added uniform maps to property-owners.

Uniform maps are only held by SceneGraph, so the API was updated to match. Event objects do not hold a copy of the map.

The maps from all 6 rendering objects are re-collected in the renderer-attachment's DoPrepareRender step if any of them have changed.

The collected map is accessible by the NewRenderer from the render thread through the UniformMapDataProvider interface, so is double buffered.

The NewRenderer maintains a list of uniform names/ program location index / property value address, and updates the uniform locations from the uniform names only when the uniform mapping has changed, not when property values change.

Change-Id: I9cec141cd7746540eda3f9049b0f4ede907b03f8
Signed-off-by: David Steele <david.steele@partner.samsung.com>