Remove compile time assert for animatable property registration
The compile time assert to check the order of the properties relies on 
the __COUNTER__ macro which generates sequential integral values starting 
from 0 at compile time and there is no way to reset it. Therefore, while 
using __COUNTER_ for both DALI_PROPERTY_REGISTRATION and 
DALI_ANIMATABLE_PROPERTY_REGISTRATION, depending on which macro is called 
first, the condition used by the assert in the other macro will always 
get the wrong count.

Change-Id: I7c263cc505050fa9bf73d652a1535fd652ec7596