Changed passing of data providers to renderers
The data providers were previously passed as a tree of pointers - this
was wrong because they weren't double buffered in the data
provider implementations.

Instead of passing as a tree, we now collate all the data providers
into a single structure before passing it to the renderer. This is
an owned object, and is doubled buffered through the message system.

This also has the advantage of being able to pass a pointer to
the DataProviders structure around rather than individual pointers.

Change-Id: I3c61fef2e8b86c8e00a1779766158b1ff2bf153b
Signed-off-by: David Steele <david.steele@partner.samsung.com>