Changed RendererAttachment to use blend flag from RenderDataProvider rather than base class's mUseBlend flag.
This fixes a issue where (when adding to the stage, removing and then re-adding) the Renderer blend flag can be become out of sync with the RenderableAttachment's mUseBlend flag.

Related: moved Renderer::mUseBlend flag and relevant methods to ImageRender since this way of controlling the blending is now specific to ImageRenderer.

Change-Id: Iebc0664946fe4c83b43291aa8004c66e004cbf22