BufferImage mirrors an external Pixel Buffer on creation.
BufferImage now will maintain its own internal copy of a buffer
and will not own any external buffer.

BufferImage will Update() from the external buffer and it is
the application's responsibility to not destroy that if an
Update() is required.

If there is no future Update() after the BufferImage is created
then the external buffer can be safely destroyed.

Change-Id: I94ca0dd6b9b3b6d883ca0ab6d3f46f75ee38f42a
Signed-off-by: Richard Underhill <r.underhill@partner.samsung.com>