BufferImage mirrors an external Pixel Buffer on creation.
BufferImage now will maintain its own internal copy of a buffer
and will not own any external buffer.

BufferImage will Update() from the external buffer and it is
the application's responsibility to not destroy that if an
Update() is required.

If there is no future Update() after the BufferImage is created
then the external buffer can be safely destroyed.

Change-Id: I7bb78d104c0c5efa24181da015b7a54c764c98ef
Signed-off-by: Richard Underhill <r.underhill@partner.samsung.com>