Implemented WeakHandle in dali-core
WeakHandle stores a weak pointer to an internal Dali object. The handle to the object
can be accessed if the object exists, and such access is not reference counted. When
the object is deleted, the weak pointer will be set to NULL, and any further attmpt
to access to a deleted object will return an empty handle.

Change-Id: I56c8b0ea398f22c1d735270e6287124bdcea740b