BufferImage mirrors an external Pixel Buffer on creation.
BufferImage now will maintain its own internal copy of a buffer
and will not own any external buffer.

BufferImage will Update() from the external buffer and it is
the application's responsibility to not destroy that if an
Update() is required.

If there is no future Update() after the BufferImage is created
then the external buffer can be safely destroyed.

Change-Id: I8e7186949ede18a5be477098667232f9261d0cfa
Signed-off-by: Richard Underhill <r.underhill@partner.samsung.com>