ResourceManager ResourcePostProcessQueue no longer double bufferred
Previously the queue consisted of two buffers which held update messages.
The resouce load message would come in but may not be stored in the current buffer.
Hence the message would only be processed on the next update when the buffers switched.

As the only post process is to notify an UPLOADED texture (DELETED is a no-op) the queue is now only of resource ID's indicating resource that has uploaded.

Change-Id: Ia38a2a32e3f45f6d3e07c0061db5cdfefa54ad12