Property::Value changed to store AngleAxis instead of Quaternion
For angles greater than 360 degrees, AngleAxis keeps it's winding
number (the number of complete rotations); whereas Quaternion loses
it. Changing Property::Value to store AngleAxis instead of Quaternion
internally enables AnimateBy to rotate 360 degrees or greater when
constructed with an AngleAxis.

Quaternions can _usually_ be converted to an Angle + Axis equivalent.
In an ideal world, there are some quaternion values that won't
convert, i.e. those for which sine(acos(.w)) is zero.

This is not an issue, as there is in practice no float value that
accurately represents N*2*PI ( for N != 0 ), consequently, the element
w generated by the Quaternion( Radian, Vector3 ) constructor is never
1 unless the initial angle was 0, which indicates no rotation.

Change-Id: I237ad5ccb1797ce7b5af9dc5e7708f65d4b785ad
Signed-off-by: David Steele <david.steele@samsung.com>